Correct version for schema.xml.
The version attribute on the schema root-node is meaningful. It
cannot be any arbitrary value. More info:
https://issues.apache.org/jira/browse/SOLR-523

The old value of '0.2' resulted in a 400 Response from Solr
when the api was accessed.

This request:
librarycloud.dev/librarycloud.v.02/api/item

Resulted in Solr throwing this exception:
SEVERE: org.apache.solr.common.SolrException: can not sort on
multivalued field: checkouts

The multivalued attribute was not introduced until Solr 1.1.
Changing the version attribute to '1.4' resolved the issue.