Rework installation guide.
Updated for clarity and to incorporate changes made since forking
from dpla/platform.git.

This could still use some more love. Seems like it will be a pain
to continue updating the documentation whenever a new version is
tagged so maybe we should consider changing up the directory
structure and punting on including the version number.

Also, it's a bit of a let down because without loading data (which
is currently "out of scope" of this guide), there's not really
much you can do with the dpla platform.